fixed types

+++ internal/pkg/comment/usecase/usecase.go
@@ -19,7 +19,7 @@ func NewUsecase(repo comment.IRepository) *Usecase {
 func (cu *Usecase) Create(input *models.InputComment, userId uint) (uint, error) {
 	comment := &models.Comment{
 		UserId: userId,
-		PinId:  input.PinId,
+		PinId:  uint(input.PinId),
 		Text:   input.Text,
 	}
